修正错误：hibernate问题

+++ project/beer/application/framework-core/src/main/java/com/apusic/ebiz/framework/core/model/IdEntity.java
@@ -180,13 +180,13 @@ public abstract class IdEntity implements Serializable {
 	private static final transient Validator validator = Validation.buildDefaultValidatorFactory().getValidator();
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	protected Long id;
+	protected Integer id;
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
 	@Override
