+++ project/beer/application/framework-core/src/test/java/com/apusic/ebiz/framework/core/dao/CrudServiceHibernateImplTest.java
@@ -92,7 +92,7 @@ public class CrudServiceHibernateImplTest {
 		// Cannot be found
 		assertNull(p);
 
-		Long id = detachedPermission.getId();
+		Integer id = detachedPermission.getId();
 		Permission atachedPermission = crudService.retrieve(Permission.class,
 				id);
 
@@ -109,7 +109,7 @@ public class CrudServiceHibernateImplTest {
 	@Transactional
 	public void updatePermission() {
 		Permission p = new Permission();
-		p.setId(1l);
+		p.setId(1);
 		p.setName("UPDATE");
 
 		crudService.update(p);
@@ -137,7 +137,7 @@ public class CrudServiceHibernateImplTest {
 	@Rollback(true)
 	public void deletePermission() {
 		Permission p2 = new Permission();
-		Long id = 20l;
+		Integer id = 20;
 		p2.setId(id);
 		crudService.delete(p2);
 	}
@@ -249,7 +249,7 @@ public class CrudServiceHibernateImplTest {
 		DummyUser user = new DummyUser();
 
 		//Exist or not exist
-		user.setId(29l);
+		user.setId(29);
 		user.setPassword("123");
 		crudService.delete(user);
 	}
